pdf thumbnails - cleanging
